backend/requirements.txt
# ==============================================================================
# File: backend/requirements.txt
# @author: Memba Co.
# ==============================================================================
# Web Framework & Server
fastapi
uvicorn[standard]
apscheduler
uvloop

# LangChain & Google AI
langchain
langchain-google-genai
google-generativeai

# Exchange & Data Analysis
ccxt
pandas
# DÜZELTME: pandas-ta ile uyumluluğu garanti etmek için numpy versiyonu sabitlendi.
numpy==1.26.4
pandas-ta

# Utilities & Telegram
python-dotenv
tenacity
requests
python-telegram-bot